requirements.txt
# Django та основні компоненти
Django==5.2.*
django-unfold==0.29.*
django-simple-history==3.4.*
django-import-export==4.0.*
django-guardian==2.4.*
django-crispy-forms==2.1.*
crispy-bootstrap5==2024.2.*
django-cors-headers==4.3.*

# Django REST Framework
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
django-filter==23.5.*

# База даних
psycopg2-binary==2.9.*

# Redis та Celery
redis==5.0.*
celery==5.3.*
django-celery-beat==2.5.*
django-celery-results==2.5.*
django-redis==5.4.*

# Конфігурація та змінні оточення
python-decouple==3.8.*
python-dotenv==1.0.*

# Генерація кодів
python-barcode==0.15.*
qrcode==7.4.*
Pillow==10.2.*

# Експорт файлів
XlsxWriter==3.1.*
reportlab==4.0.*

# HTTP клієнт та асинхронність
requests==2.31.*
aiohttp==3.9.*
tenacity==8.2.*

# Системна інформація (для агента)
psutil==5.9.*

# Дата та час
python-dateutil==2.8.*

# Валідація та парсинг
marshmallow==3.20.*

# Тестування (додатково)
pytest==7.4.*
pytest-django==4.8.*
pytest-cov==4.1.*
factory-boy==3.3.*

# Лінтинг та форматування (додатково)
black==23.12.*
flake8==7.0.*
isort==5.13.*

# Моніторинг помилок (додатково)
sentry-sdk==1.40.*

# API документація (додатково)
drf-spectacular==0.27.*

# Безпека
django-security==0.17.*

# Двофакторна аутентифікація
pyotp==2.9.*
qrcode[pil]==7.4.*

# Rate limiting
django-ratelimit==4.1.*

# WSGI сервер для продакшену
gunicorn==21.2.*
whitenoise==6.6.*